* commit 'e959bdc036035496ab694fe6ba225fe0917f6497': (429 commits)
  Remove extraneous logger.info and change a warning to logger.warning
  Make the prereq test more flexible and correct
  Implement a much simpler script_regression_test tests
  Partial implementation of the prereq test. Still need to find out a generic way of verifying that the dependant job actually depends on the other one
  Revert change from afterok to afterany; this should only be done if the user specifically requests it
  Implements the prereq argument for case.submit, allowing the user to specify jobs which should finish (not necessarily successfully) before running the current job
  Update ChangeLog
  Change case.st_archive dependency string to not include strings indicating unsupported logic. Fix cobalt depend_string. Add depend_separator field to support LSF
  update mpt and pnetcdf on cheyenne
  Update description for TOTALPES
  Include spare nodes in COST_PES
  Update ChangeLog
  Add mapping files for T42_T42_mg17
  Significant refactor to aprun
  Fix aprun unit test.
  Fix COST_PES and TOTALPES for titan (aprun)
  Fix for multiple couplers. Also fix typo regression and simplify code somewhat
  update comment
  empty proxy tag in machines file invalid, removed from anlworkstation
  response to review
  ...

+++ scripts/lib/CIME/SystemTests/erio.py
@@ -24,7 +24,7 @@ class ERIO(SystemTestsCommon):
         expect(self._stop_n > 0, "Bad STOP_N: {:d}".format(self._stop_n))
 
         # Move to config_tests.xml once that's ready
-        rest_n = self._stop_n/2 + 1
+        rest_n = int(self._stop_n/2) + 1
         self._case.set_value("REST_N", rest_n)
         self._case.set_value("REST_OPTION", stop_option)
         self._case.set_value("HIST_N", self._stop_n)
@@ -39,7 +39,7 @@ class ERIO(SystemTestsCommon):
     def _restart_run(self, pio_type, other_pio_type):
         stop_option = self._case.get_value("STOP_OPTION")
 
-        rest_n = self._stop_n/2 + 1
+        rest_n = int(self._stop_n/2) + 1
         stop_new = self._stop_n - rest_n
         expect(stop_new > 0, "ERROR: stop_n value {:d} too short {:d} {:d}".format(stop_new,self._stop_n,rest_n))
