+++ scripts/lib/CIME/SystemTests/pet.py
@@ -49,7 +49,7 @@ class PET(SystemTestsCompareTwo):
         # machines, if the mpiexec tries to exceed the procs-per-node that were given
         # to the batch submission, things break. Setting MAX_TASKS_PER_NODE to half of
         # it original value prevents this.
-        self._case.set_value("MAX_TASKS_PER_NODE", self._case.get_value("MAX_TASKS_PER_NODE") / 2)
+        self._case.set_value("MAX_TASKS_PER_NODE", int(self._case.get_value("MAX_TASKS_PER_NODE") / 2))
 
         # Need to redo case_setup because we may have changed the number of threads
         case_setup(self._case, reset=True)
